pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autocommitmessage"
version = "0.1.2"
authors = [
  { name="SLPRYSQUID", email="coleseto123@gmail.com" },
]
description = "A tool to automatically generate git commit messages using an LLM."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "GitPython>=3.1.45",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langsmith>=0.4.27",
    "python-dotenv>=0.9.9",
]

[project.urls]
"Homepage" = "https://github.com/slipperysquid/AutoCommitMessage" 
"Bug Tracker" = "https://github.com/slipperysquid/AutoCommitMessage/issues"

[project.scripts]
autocommitmessage = "autocommitmessage.commit_generator:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-mock>=3.15.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["autocommitmessage*"]

[tool.pytest.ini_options]
markers = [
    "e2e: marks tests as end-to-end tests (makes real API calls)",
    "llm: marks tests as LLM tests",
]